Add network interface monitoring to telegraf configuration
Using the ansible ini_file module I wasn't able to get it to uncomment
the section just by asking it to be present. Explicitly specifying an
option and a value causes it to be written out. Since we don't know
a priori what the enames of the interfaces will be we use "*" here.
This will also include the loopback lo0, but that will liekly be in
the noise compared to the others.

Fixes #4