Other fixes for Oracle. Security updates break the install (we need to find a way to run them, as disabling them is a temporary "fix").
Also the GPG key check for slurm is broken. This is removed,
but this is also just a temporary "fix". We should find a way of
packaging up what we need - maybe download the slurm RPM via
the git clone of the ansible repo? I guess this is a longer term
thing as we think about what the foundation OS layer should be,
and how we install software like slurm on top of that layer.